Refactored class structure
