+++ app/src/main/kotlin/com/exactpro/th2/lwdataprovider/SseEvent.kt
@@ -19,15 +19,15 @@ package com.exactpro.th2.lwdataprovider
 import com.exactpro.cradle.Direction
 import com.exactpro.cradle.messages.StoredMessageId
 import com.exactpro.th2.lwdataprovider.SseEvent.Companion.DATA_CHARSET
-import com.exactpro.th2.lwdataprovider.entities.responses.ByteBufferPool
+import com.exactpro.th2.lwdataprovider.entities.responses.ser.ByteBufferPool
 import com.exactpro.th2.lwdataprovider.entities.responses.LastScannedObjectInfo
 import com.exactpro.th2.lwdataprovider.entities.responses.LwEvent
 import com.exactpro.th2.lwdataprovider.entities.responses.PageInfo
 import com.exactpro.th2.lwdataprovider.entities.responses.ProviderMessage53
 import com.exactpro.th2.lwdataprovider.entities.responses.ProviderMessage53Transport
 import com.exactpro.th2.lwdataprovider.entities.responses.ResponseMessage
-import com.exactpro.th2.lwdataprovider.entities.responses.serialize
-import com.exactpro.th2.lwdataprovider.entities.responses.serializeJsonData
+import com.exactpro.th2.lwdataprovider.entities.responses.ser.serialize
+import com.exactpro.th2.lwdataprovider.entities.responses.ser.serializeJsonData
 import com.fasterxml.jackson.databind.ObjectMapper
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerializationStrategy
